RIP-552, alert w/o redirect when failed dev-auth
