+++ src/api/auth.ts
@@ -4,8 +4,8 @@ export function becomeUser(canvasSiteId: number, uid: string) {
   return utils.post('/api/auth/become_user', {canvasSiteId, uid}, true)
 }
 
-export function devAuthLogIn(canvasSiteId: number, uid: string, password: string) {
-  return utils.post('/api/auth/dev_auth', {canvasSiteId, password, uid}, true)
+export function devAuthLogIn(canvasSiteId: string, uid: string, password: string) {
+  return utils.post('/api/auth/dev_auth', {canvasSiteId, password, uid}, false)
 }
 
 export function getCasLoginURL() {
