+++ src/utils.ts
@@ -2,6 +2,8 @@ import {capitalize, concat, get, head, includes, initial, join, last, noop, star
 import {nextTick} from 'vue'
 import {useContextStore} from '@/stores/context'
 
+const SKIP_REDIRECT_ON_ERROR = ['/api/auth/dev_auth']
+
 export const isInIframe = !!window.parent.frames.length
 
 export function decamelize(str: string, separator=' ') {
@@ -63,7 +65,9 @@ export function initializeAxios(app: any, axios: any) {
         return axios.get(`${apiBaseUrl}/api/user/my_profile`).then((data: any) => {
           const currentUser = data
           useContextStore().setCurrentUser(currentUser)
-          if (!currentUser.isAuthenticated) {
+          const apiUrl = trim(get(error.config, 'url'))
+          const skipRedirect = currentUser.isAuthenticated || SKIP_REDIRECT_ON_ERROR.some(apiPath => apiUrl.includes(apiPath))
+          if (!skipRedirect) {
             useContextStore().setApplicationState(errorStatus, 'Your session has expired')
           }
           return Promise.reject(error)
